New file with common initialization part
