+++ projects/postgresql/fuzzer/json_parser_fuzzer.c
@@ -16,95 +16,13 @@
 
 #include "postgres.h"
 
-#include "access/xlog.h"
-#include "access/xact.h"
 #include "common/jsonapi.h"
-#include "common/username.h"
-#include "executor/spi.h"
-#include "jit/jit.h"
-#include "libpq/libpq.h"
-#include "libpq/pqsignal.h"
 #include "mb/pg_wchar.h"
-#include "miscadmin.h"
-#include "optimizer/optimizer.h"
-#include "parser/analyze.h"
-#include "parser/parser.h"
-#include "storage/proc.h"
-#include "tcop/tcopprot.h"
-#include "utils/datetime.h"
-#include "utils/memdebug.h"
 #include "utils/memutils.h"
-#include "utils/portal.h"
-#include "utils/snapmgr.h"
-#include "utils/timeout.h"
-
-const char *progname = "progname";
-static const char *userDoption;
-static MemoryContext row_description_context = NULL;
-static StringInfoData row_description_buf;
-static const char *dbname = NULL;
-static const char *username = NULL;
-
-static void fuzzer_exit(){
-  if(!username)
-    pfree((void *) username);
-}
-
+#include "utils/memdebug.h"
 
 int __attribute__((constructor)) Initialize(void) {
-  int argc = 4;
-  char *argv[4];
-  argv[0] = "tmp_install/usr/local/pgsql/bin/postgres";
-  argv[1] = "-D\"/tmp/json_db/data\"";
-  argv[2] = "-F";
-  argv[3] = "-k\"/tmp/pg_dbfuzz\"";
-
-  system("tar -xvf json_db.tar.gz -C /tmp/");
-  
-  progname = get_progname(argv[0]);
-  MemoryContextInit();
-
-  username = strdup(get_user_name_or_exit(progname));
-	 
-  InitStandaloneProcess(argv[0]);
-  SetProcessingMode(InitProcessing);
-  InitializeGUCOptions();
-  process_postgres_switches(argc, argv, PGC_POSTMASTER, &dbname);
-  dbname = "dbfuzz";
-
-  userDoption = "/tmp/json_db/data";
-  SelectConfigFiles(userDoption, progname);
-
-  checkDataDir();
-  ChangeToDataDir();
-  CreateDataDirLockFile(false);
-  LocalProcessControlFile(false);
-  InitializeMaxBackends();
-		 
-  BaseInit();
-  InitProcess();
-  PG_SETMASK(&UnBlockSig);
-  InitPostgres(dbname, InvalidOid, username, InvalidOid, NULL, false);
- 
-  SetProcessingMode(NormalProcessing);
-
-  BeginReportingGUCOptions();
-  process_session_preload_libraries();
-
-  MessageContext = AllocSetContextCreate(TopMemoryContext,
-										 "MessageContext",
-										 ALLOCSET_DEFAULT_SIZES);
-  row_description_context = AllocSetContextCreate(TopMemoryContext,
-												  "RowDescriptionContext",
-												  ALLOCSET_DEFAULT_SIZES);
-  MemoryContextSwitchTo(row_description_context);
-  initStringInfo(&row_description_buf);
-  MemoryContextSwitchTo(TopMemoryContext);
-
-  PgStartTime = GetCurrentTimestamp();
-  whereToSendOutput = DestNone;
-  Log_destination = 0;
-  atexit(fuzzer_exit);
+  FuzzerInitialize("json_db");
   return 0;
 }
 
@@ -122,7 +40,6 @@ int LLVMFuzzerTestOneInput(const uint8_t* data, size_t size) {
 	memcpy(buffer, data, size);
 
 	MemoryContextInit();
-	set_stack_base();
 	sem = nullSemAction;
 	lex = makeJsonLexContextCstringLen(buffer, size+1, PG_UTF8, true);
