+++ projects/postgresql/fuzzer/simple_query_fuzzer.c
@@ -35,12 +35,6 @@
 #include "utils/snapmgr.h"
 #include "utils/timeout.h"
 
-const char *progname;
-static const char *userDoption;
-static MemoryContext row_description_context = NULL;
-static StringInfoData row_description_buf;
-static const char *dbname = NULL;
-static const char *username = NULL;
 
 static void
 exec_simple_query(const char *query_string)
@@ -99,66 +93,9 @@ exec_simple_query(const char *query_string)
     }
 }
 
-static void fuzzer_exit(){
-  if(!username)
-    pfree((void *) username);
-}
-
 
 int __attribute__((constructor)) Initialize(void) {
-  int argc = 4;
-  char *argv[4];
-  argv[0] = "tmp_install/usr/local/pgsql/bin/postgres";
-  argv[1] = "-D\"/tmp/query_db/data\"";
-  argv[2] = "-F";
-  argv[3] = "-k\"/tmp/pg_dbfuzz\"";
-
-  system("tar -xvf query_db.tar.gz -C /tmp/");
-  
-  progname = get_progname(argv[0]);
-  MemoryContextInit();
-
-  username = strdup(get_user_name_or_exit(progname));
-	 
-  InitStandaloneProcess(argv[0]);
-  SetProcessingMode(InitProcessing);
-  InitializeGUCOptions();
-  process_postgres_switches(argc, argv, PGC_POSTMASTER, &dbname);
-  dbname = "dbfuzz";
-
-  userDoption = "/tmp/query_db/data";
-  SelectConfigFiles(userDoption, progname);
-
-  checkDataDir();
-  ChangeToDataDir();
-  CreateDataDirLockFile(false);
-  LocalProcessControlFile(false);
-  InitializeMaxBackends();
-		 
-  BaseInit();
-  InitProcess();
-  PG_SETMASK(&UnBlockSig);
-  InitPostgres(dbname, InvalidOid, username, InvalidOid, NULL, false);
- 
-  SetProcessingMode(NormalProcessing);
-
-  BeginReportingGUCOptions();
-  process_session_preload_libraries();
-
-  MessageContext = AllocSetContextCreate(TopMemoryContext,
-					 "MessageContext",
-					 ALLOCSET_DEFAULT_SIZES);
-  row_description_context = AllocSetContextCreate(TopMemoryContext,
-						  "RowDescriptionContext",
-						  ALLOCSET_DEFAULT_SIZES);
-  MemoryContextSwitchTo(row_description_context);
-  initStringInfo(&row_description_buf);
-  MemoryContextSwitchTo(TopMemoryContext);
-
-  PgStartTime = GetCurrentTimestamp();
-  whereToSendOutput = DestNone;
-  Log_destination = 0;
-  atexit(fuzzer_exit);
+  FuzzerInitialize("query_db");
   return 0;
 }
